Update DondeDine.py
Was able to add the ability to get inputs from two receivers and labeling them boyfriend and girlfriend. However, I would like to be able to all the user to input their name or any creative name they would like in the future.

I also created a string function that can be found in the widget.