Random Button + User Customization + Branding Update DondeDine.py
I have added a randomization button and scrapped the revoke option as it was to complicated and did not work. I have also updated the branding to match the logo I made but am trying to figure out how to upload a logo onto the tkinter model. Also added a hover color feature that matches the aesthetic as well.

A big improvement, however, was my ability to change the user customization and to update live when applying the name and adding it to the titles.